refactor: relocate moves dex
