+++ pokemon/dex/__init__.py
@@ -21,7 +21,7 @@ BASE_PATH = Path(__file__).resolve().parents[2]
 # Raw data files
 INPUT_PATH = BASE_PATH / "helpers" / "scripts" / "input"
 POKEDEX_PATH = BASE_PATH / "pokemon" / "dex" / "pokedex" / "__init__.py"
-MOVEDEX_PATH = BASE_PATH / "pokemon" / "data" / "moves.py"
+MOVEDEX_PATH = BASE_PATH / "pokemon" / "dex" / "moves" / "movesdex.py"
 ABILITYDEX_PATH = BASE_PATH / "pokemon" / "dex" / "abilities" / "abilitiesdex.py"
 ITEMDEX_PATH = BASE_PATH / "pokemon" / "dex" / "items" / "itemsdex.py"
 CONDITIONDEX_PATH = BASE_PATH / "pokemon" / "dex" / "conditions.py"
